Created new playlist route and class, with barebones meta population.
