+++ server/app/models/__init__.py
@@ -1,3 +1,3 @@
-from .track import Track
-from .playlist import Playlist
+from .old.track import Track
+from .old.playlist import Playlist
 from .user import User
